FIX: Separando os endpoints por pastas de acordo com as classes
