+++ src/endpoints/Turmas/criarTurma.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { connection } from "../data/connection";
+import { connection } from "../../data/connection";
 
 export default async function criarTurma(
     req: Request,
